Added images Timer not working

+++ app/src/main/java/com/hkngtech/continuoustimer/others/Constants.kt
@@ -3,9 +3,9 @@ package com.hkngtech.continuoustimer.others
 object Constants {
 
     fun miltotime(mil : Long) : String{
-        val seconds = ((mil.toLong() / 1000) % 60).toInt()
-        val minutes = ((mil.toLong() / (1000 * 60) % 60)).toInt()
-        val hours = ((mil.toLong() / (1000 * 60 * 60) % 24)).toInt()
+        val seconds = ((mil / 1000) % 60).toInt()
+        val minutes = ((mil / (1000 * 60) % 60)).toInt()
+        val hours = ((mil / (1000 * 60 * 60) % 24)).toInt()
         return "$hours:$minutes:$seconds"
     }
 }
